requirements.txt
streamlit
requests
beautifulsoup4
gspread
google-auth
